Changes to animation BP and master grunt type
added all I can in terms of movement and the aim offset works just implementing into the enemies AI with searching and stuff is a bit beyond me. But movement works the strafing isnt the best blend, but due to it only happening when the enemy is engaged it combat it should be okay. The enemy blueprint in both Master and Tech ive changed to the enemy and its animation BP so you dont need to change that either